remove phi instruction and replace old terminal instructions with new one
