+++ crates/ir/src/lib.rs
@@ -47,9 +47,7 @@ pub mod instr {
         BinaryFloatInstr,
         BinaryIntInstr,
         BranchInstr,
-        BranchInstr2,
         BranchTableInstr,
-        BranchTableInstr2,
         CallIndirectInstr,
         CallInstr,
         CompareFloatInstr,
@@ -61,27 +59,21 @@ pub mod instr {
         FloatToIntInstr,
         HeapAddrInstr,
         IfThenElseInstr,
-        IfThenElseInstr2,
         Instruction,
         IntInstr,
         IntToFloatInstr,
         LoadInstr,
         MemoryGrowInstr,
         MemorySizeInstr,
-        PhiInstr,
         PromoteFloatInstr,
         ReinterpretInstr,
         ReturnInstr,
-        ReturnInstr2,
         SelectInstr,
         ShiftIntInstr,
         StoreInstr,
         TailCallIndirectInstr,
-        TailCallIndirectInstr2,
         TailCallInstr,
-        TailCallInstr2,
         TerminalInstr,
-        TerminalInstr2,
         TruncateIntInstr,
         UnaryFloatInstr,
         UnaryIntInstr,
